feat: fix design issues
