+++ packages/web/src/features/proposal/components/ViewerExpenditureTable.tsx
@@ -7,38 +7,45 @@ import {
   getCoreRowModel,
   useReactTable,
 } from "@tanstack/react-table";
-import Table from "@sparcs-students/web/common/components/Table";
-import LightTag from "@sparcs-students/web/common/components/Tag/LightTag";
-import { getTagDetail } from "@sparcs-students/web/utils/getTagDetail";
+import Table from "@sparcs-students/web/common/components/Table/Table";
 import {
-  budgetDivisionIncomeTagList,
-  budgetDomainTagList,
-} from "@sparcs-students/web/constants/tableTagList";
-import {
+  BudgetClassExpenseE,
   BudgetDivisionIncomeE,
   BudgetDomainE,
 } from "@sparcs-students/interface/common/enum/budget.enum";
+import LightTag from "@sparcs-students/web/common/components/Tag/LightTag";
+import {
+  getDarkTagDetail,
+  getTagDetail,
+} from "@sparcs-students/web/utils/getTagDetail";
+import {
+  budgetClassExpenseTagList,
+  budgetDivisionIncomeTagList,
+  budgetDomainTagList,
+} from "@sparcs-students/web/constants/tableTagList";
+import { useFormatter } from "next-intl";
 import DetailButton from "@sparcs-students/web/features/proposal/components/_atomic/DetailButton";
 import DarkTag from "@sparcs-students/web/common/components/Tag/DarkTag";
+import ExpenditureHelpButton from "./_atomic/ExpenditureHelpButton";
 
-export interface IncomeProps {
+export interface ExpenditureProps {
   code: number;
   budgetDomain: BudgetDomainE;
   budgetDivisionIncome: BudgetDivisionIncomeE;
-  item: string;
+  name: string;
+  item: BudgetClassExpenseE;
   lastYear: number;
   thisYear: number;
   ratio: number;
   reason: string;
   status: string;
-  explanation: string;
 }
 
-interface IncomeTableProps {
-  data: IncomeProps[];
+interface ExpenditureTableProps {
+  data: ExpenditureProps[];
 }
 
-const columnHelper = createColumnHelper<IncomeProps>();
+const columnHelper = createColumnHelper<ExpenditureProps>();
 
 const columns = [
   columnHelper.accessor("code", {
@@ -85,21 +92,33 @@ const columns = [
     },
     size: 120,
   }),
+  columnHelper.accessor("name", {
+    id: "name",
+    header: "사업명",
+    cell: info => info.getValue(),
+    size: 80,
+  }),
   columnHelper.accessor("item", {
     id: "item",
     header: "항목",
-    cell: info => info.getValue(),
-    size: 180,
+    cell: info => {
+      const { color, text } = getDarkTagDetail(
+        info.getValue(),
+        budgetClassExpenseTagList,
+      );
+      return <DarkTag color={color}>{text}</DarkTag>;
+    },
+    size: 100,
   }),
   columnHelper.accessor("lastYear", {
     id: "lastYear",
     header: "작년 결산",
     cell: info => {
-      const formatter = new Intl.NumberFormat("ko-KR", {
+      const format = useFormatter();
+      return format.number(info.getValue(), {
         style: "currency",
         currency: "KRW",
       });
-      return formatter.format(info.getValue());
     },
     size: 120,
   }),
@@ -107,11 +126,11 @@ const columns = [
     id: "thisYear",
     header: "올해 예산",
     cell: info => {
-      const formatter = new Intl.NumberFormat("ko-KR", {
+      const format = useFormatter();
+      return format.number(info.getValue(), {
         style: "currency",
         currency: "KRW",
       });
-      return formatter.format(info.getValue());
     },
     size: 120,
   }),
@@ -157,15 +176,9 @@ const columns = [
     // TODO: Add LightTag by enum
     size: 90,
   }),
-  columnHelper.accessor("explanation", {
-    id: "explanation",
-    header: "설명",
-    cell: info => <DetailButton detail={info.getValue()} />,
-    size: 60,
-  }),
 ];
 
-const IncomeTable: React.FC<IncomeTableProps> = ({ data }) => {
+const ViewerExpenditureTable: React.FC<ExpenditureTableProps> = ({ data }) => {
   const [loaded, setLoaded] = useState(false);
 
   useEffect(() => {
@@ -181,12 +194,15 @@ const IncomeTable: React.FC<IncomeTableProps> = ({ data }) => {
 
   return (
     <FlexWrapper direction="column" gap={16}>
-      <Typography fs={24} lh={30} color="BLACK" fw="SEMIBOLD">
-        수입
-      </Typography>
+      <FlexWrapper direction="row" gap={12}>
+        <Typography fs={24} lh={30} color="BLACK" fw="SEMIBOLD">
+          지출
+        </Typography>
+        <ExpenditureHelpButton />
+      </FlexWrapper>
       {loaded && <Table table={table} />}
     </FlexWrapper>
   );
 };
 
-export default IncomeTable;
+export default ViewerExpenditureTable;
